gwpy.testing: new fixtures module
to hold general-purpose fixtures that previously were in sub-package tests
