+++ gwpy/conftest.py
@@ -26,11 +26,13 @@ import numpy
 
 from matplotlib import use
 
-# ignore errors due from pyplot.show() using Agg
-warnings.filterwarnings('ignore', message=".*non-GUI backend.*")
+# register custom fixtures for all test modules
+from gwpy.testing.fixtures import *
 
 # set random seed to 1 for reproducability
 numpy.random.seed(1)
 
+# ignore errors due from pyplot.show() using Agg
+warnings.filterwarnings('ignore', message=".*non-GUI backend.*")
 # force Agg for all tests
 use('agg', warn=False, force=True)
