+++ gwpy/plot/tests/test_axes.py
@@ -31,8 +31,7 @@ from ...time import LIGOTimeGPS
 from ...types import (Series, Array2D)
 from ...testing import utils
 from .. import Axes
-from .utils import (usetex,  # pylint: disable=unused-import
-                    AxesTestBase)
+from .utils import AxesTestBase
 
 numpy.random.seed(0)
